refactor: api 호출 커스텀 훅 제거 #54
